backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
python-multipart
pydantic
chromadb
sentence-transformers
datasets
face-recognition
numpy
pillow
pydantic-settings